airflow/local-executor/requirements.txt
pandas
requests
python-dotenv
sqlalchemy
pymysql
numpy